Removed/changed to only include things in min implementation
The minimum implementation will focus mostly on things that I in 326
dealt with. It is a proof of concept that implementing iTrust in Python
is feasible without having too much of an unfinished product.

If I have the time, I may go back and do the other parts that were also
in iTrust, but I want to focus on the minimal implementation that
surronds what we did.